pyproject.toml
[tool.poetry]
name = "tmodloader_mod_manager"
version = "0.1.0"
description = ""
authors = ["Bobby R. Ward <bobbyrward@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pysteamcmd = "^0.3.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tmodloader_mod_manager = "tmodloader_mod_manager.main:main"
